refactor(cli): put write functions in a separate module
